wip: create osd drawer service
