+++ libs/vre/shared/app-representations/src/lib/still-image/still-image.component.ts
@@ -1,41 +1,27 @@
 import {
   ChangeDetectionStrategy,
-  ChangeDetectorRef,
   Component,
   ElementRef,
-  Inject,
   Input,
   OnChanges,
   OnDestroy,
   OnInit,
-  Renderer2,
   SimpleChanges,
 } from '@angular/core';
-import { MatDialog } from '@angular/material/dialog';
 import { MatIconRegistry } from '@angular/material/icon';
 import { DomSanitizer } from '@angular/platform-browser';
-import {
-  Constants,
-  KnoraApiConnection,
-  Point2D,
-  ReadColorValue,
-  ReadProject,
-  ReadResource,
-  ReadStillImageFileValue,
-  RegionGeometry,
-} from '@dasch-swiss/dsp-js';
+import { Constants, ReadProject, ReadResource, ReadStillImageFileValue } from '@dasch-swiss/dsp-js';
 import { ReadStillImageExternalFileValue } from '@dasch-swiss/dsp-js/src/models/v2/resources/values/read/read-file-value';
-import { DspApiConnectionToken } from '@dasch-swiss/vre/shared/app-config';
 import { AppError } from '@dasch-swiss/vre/shared/app-error-handler';
 import * as OpenSeadragon from 'openseadragon';
-import { combineLatest, Subject } from 'rxjs';
-import { distinctUntilChanged, filter, switchMap, take, takeUntil } from 'rxjs/operators';
-import { AddRegionFormDialogComponent, AddRegionFormDialogProps } from '../add-region-form-dialog.component';
+import { Subject } from 'rxjs';
+import { take, takeUntil } from 'rxjs/operators';
 import { RegionService } from '../region.service';
 import { RepresentationService } from '../representation.service';
 import { FileInfo } from '../representation.types';
 import { ResourceFetcherService } from '../resource-fetcher.service';
 import { IIIFUrl } from '../third-party-iiif/third-party-iiif';
+import { OsdDrawerService } from './osd-drawer.service';
 import { osdViewerConfig } from './osd-viewer.config';
 import { StillImageHelper } from './still-image-helper';
 
@@ -48,6 +34,7 @@ export interface PolygonsForRegion {
   templateUrl: './still-image.component.html',
   styleUrls: ['./still-image.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
+  providers: [OsdDrawerService],
 })
 export class StillImageComponent implements OnInit, OnChanges, OnDestroy {
   @Input({ required: true }) resource!: ReadResource;
@@ -57,11 +44,8 @@ export class StillImageComponent implements OnInit, OnChanges, OnDestroy {
   isPng: boolean = false;
   imageFileInfo: FileInfo | undefined = undefined;
   failedToLoad = false;
-  regionDrawMode = false; // stores whether viewer is currently drawing a region
 
-  private _regionDragInfo: { overlayElement: HTMLElement; startPos: OpenSeadragon.Point } | null = null; // stores the information of the first click for drawing a region
   private _viewer: OpenSeadragon.Viewer | undefined;
-  private _regions: PolygonsForRegion = {};
   private _ngUnsubscribe = new Subject<void>();
 
   get imageFileValue(): ReadStillImageFileValue | ReadStillImageExternalFileValue | undefined {
@@ -74,17 +58,13 @@ export class StillImageComponent implements OnInit, OnChanges, OnDestroy {
   }
 
   constructor(
-    @Inject(DspApiConnectionToken)
-    private _dspApiConnection: KnoraApiConnection,
-    private _dialog: MatDialog,
     private _domSanitizer: DomSanitizer,
     private _elementRef: ElementRef,
     private _matIconRegistry: MatIconRegistry,
-    private _renderer: Renderer2,
     private _rs: RepresentationService,
     private _regionService: RegionService,
     private _resourceFetcherService: ResourceFetcherService,
-    private _cd: ChangeDetectorRef
+    public _osdDrawerService: OsdDrawerService
   ) {
     OpenSeadragon.setString('Tooltips.Home', '');
     OpenSeadragon.setString('Tooltips.ZoomIn', '');
@@ -100,31 +80,9 @@ export class StillImageComponent implements OnInit, OnChanges, OnDestroy {
 
   ngOnInit() {
     this._setupViewer();
+    this._osdDrawerService.onInit(this._viewer!);
+    this._osdDrawerService.addRegionDrawer();
     this._loadImages();
-
-    this._regionService.imageIsLoaded$
-      .pipe(
-        filter(loaded => loaded),
-        switchMap(() =>
-          combineLatest([this._regionService.showRegions$.pipe(distinctUntilChanged()), this._regionService.regions$])
-        )
-      )
-      .pipe(filter(() => !!this._viewer)) // only proceed if the viewer is already set up
-      .subscribe(([showRegion, regions]) => {
-        this._removeOverlays();
-
-        if (showRegion) {
-          this._renderRegions();
-        }
-      });
-
-    this._regionService.highlightRegion$.subscribe(region => {
-      if (region === null) {
-        this._unhighlightAllRegions();
-        return;
-      }
-      this._highlightRegion(region);
-    });
 
     this._resourceFetcherService.settings.imageFormatIsPng
       .asObservable()
@@ -151,155 +109,6 @@ export class StillImageComponent implements OnInit, OnChanges, OnDestroy {
     this._ngUnsubscribe.complete();
   }
 
-  resetStatus() {
-    this.defaultFailureStatus = 404;
-  }
-
-  setForbiddenStatus() {
-    this.defaultFailureStatus = 403;
-    this._onFailedImageLoad();
-  }
-
-  /**
-   * opens the dialog to enter further properties for the region after it has been drawn and calls the function to upload the region after confirmation
-   * @param startPoint the start point of the drawing
-   * @param endPoint the end point of the drawing
-   * @param imageSize the image size for calculations
-   * @param overlay the overlay element that represents the region
-   */
-  private _openRegionDialog(startPoint: Point2D, endPoint: Point2D, imageSize: Point2D, overlay: Element): void {
-    if (!this._viewer) {
-      return;
-    }
-    this._dialog
-      .open<AddRegionFormDialogComponent, AddRegionFormDialogProps>(AddRegionFormDialogComponent, {
-        data: {
-          resourceIri: this.resource.id,
-        },
-      })
-      .afterClosed()
-      .subscribe(data => {
-        this._viewer!.removeOverlay(overlay);
-        if (data) {
-          this._uploadRegion(startPoint, endPoint, imageSize, data.color, data.comment, data.label);
-        }
-      });
-  }
-
-  /**
-   * uploads the region after being prepared by the dialog
-   * @param startPoint the start point of the drawing
-   * @param endPoint the end point of the drawing
-   * @param imageSize the image size for calculations
-   * @param color the value for the color entered in the form
-   * @param comment the value for the comment entered in the form
-   * @param label the value for the label entered in the form
-   */
-  private _uploadRegion(
-    startPoint: Point2D,
-    endPoint: Point2D,
-    imageSize: Point2D,
-    color: string,
-    comment: string,
-    label: string
-  ) {
-    this._dspApiConnection.v2.res
-      .createResource(
-        StillImageHelper.getPayloadUploadRegion(
-          this.resource.id,
-          this.resource.attachedToProject,
-          startPoint,
-          endPoint,
-          imageSize,
-          color,
-          comment,
-          label
-        )
-      )
-      .subscribe(res => {
-        this._viewer?.destroy(); // canvas-click event doesnt work anymore, is it viewer bug?
-        this._setupViewer();
-        this._loadImages();
-
-        const regionId = (res as ReadResource).id;
-        this._regionService.updateRegions();
-        this._regionService.highlightRegion(regionId);
-        this._regionService.showRegions(true);
-      });
-  }
-
-  /**
-   * set up function for the region drawer
-   */
-  private _addRegionDrawer(): void {
-    const viewer = this._assertOSDViewer();
-    // eslint-disable-next-line no-new
-    new OpenSeadragon.MouseTracker({
-      element: viewer.canvas,
-      pressHandler: event => {
-        if (!this.regionDrawMode) {
-          return;
-        }
-        const overlayElement: HTMLElement = this._renderer.createElement('div');
-        overlayElement.style.background = 'rgba(255,0,0,0.3)';
-        const viewportPos = viewer.viewport.pointFromPixel((event as OpenSeadragon.ViewerEvent).position!);
-        viewer.addOverlay(overlayElement, new OpenSeadragon.Rect(viewportPos.x, viewportPos.y, 0, 0));
-        this._regionDragInfo = {
-          overlayElement,
-          startPos: viewportPos,
-        };
-      },
-      dragHandler: event => {
-        if (!this._regionDragInfo) {
-          return;
-        }
-        const viewPortPos = viewer.viewport.pointFromPixel((event as OpenSeadragon.ViewerEvent).position!);
-        const diffX = viewPortPos.x - this._regionDragInfo.startPos.x;
-        const diffY = viewPortPos.y - this._regionDragInfo.startPos.y;
-        const location = new OpenSeadragon.Rect(
-          Math.min(this._regionDragInfo.startPos.x, this._regionDragInfo.startPos.x + diffX),
-          Math.min(this._regionDragInfo.startPos.y, this._regionDragInfo.startPos.y + diffY),
-          Math.abs(diffX),
-          Math.abs(diffY)
-        );
-
-        viewer.updateOverlay(this._regionDragInfo.overlayElement, location);
-        this._regionDragInfo.endPos = viewPortPos;
-      },
-      releaseHandler: () => {
-        if (this.regionDrawMode) {
-          const imageSize = viewer.world.getItemAt(0).getContentSize();
-          const startPoint = viewer.viewport.viewportToImageCoordinates(this._regionDragInfo.startPos);
-          const endPoint = viewer.viewport.viewportToImageCoordinates(this._regionDragInfo.endPos);
-          this._openRegionDialog(startPoint, endPoint, imageSize, this._regionDragInfo.overlayElement);
-          this._regionDragInfo = null;
-          this.regionDrawMode = false;
-          viewer.setMouseNavEnabled(true);
-        }
-      },
-    });
-  }
-
-  private _highlightRegion(regionIri: string) {
-    const activeRegions: HTMLElement[] = this._regions[regionIri];
-
-    if (activeRegions !== undefined) {
-      for (const pol of activeRegions) {
-        pol.setAttribute('class', 'region active');
-      }
-    }
-  }
-
-  private _unhighlightAllRegions() {
-    for (const reg in this._regions) {
-      if (reg in this._regions) {
-        for (const pol of this._regions[reg]) {
-          pol.setAttribute('class', 'region');
-        }
-      }
-    }
-  }
-
   private _setupViewer(): void {
     const viewerContainer = this._elementRef.nativeElement.getElementsByClassName('osd-container')[0];
 
@@ -315,21 +124,15 @@ export class StillImageComponent implements OnInit, OnChanges, OnDestroy {
         viewerContainer.classList.remove('fullscreen');
       }
     });
-
-    this._addRegionDrawer();
   }
 
   private _assertOSDViewer(): OpenSeadragon.Viewer {
     if (!this._viewer) {
-      throw new AppError('OpenSeadragon Viewer not initialized');
+      throw new AppError('OpenSeaDragon Viewer not initialized');
     }
     return this._viewer;
   }
 
-  /**
-   * adds all images in this.images to the _viewer.
-   * Images are positioned in a horizontal row next to each other.
-   */
   private _openInternalImages(): void {
     const viewer = this._assertOSDViewer();
     this.failedToLoad = false;
@@ -339,11 +142,13 @@ export class StillImageComponent implements OnInit, OnChanges, OnDestroy {
     // display only the defined range of this.images
     const tileSources: object[] = StillImageHelper.prepareTileSourcesFromFileValues(fileValues, this.isPng);
     viewer.addOnceHandler('open', args => {
+      console.log('arg', args);
       // check if the current image exists
-      if (
-        this.imageFileValue instanceof ReadStillImageFileValue &&
-        this.imageFileValue.fileUrl.includes(args.source['id'])
-      ) {
+      if (this.imageFileValue instanceof ReadStillImageFileValue) {
+        /** TODO there was this weird condition inside
+                 this.imageFileValue.fileUrl.includes(args.source!['id']!)
+                 */
+
         // enable mouse navigation incl. zoom
         viewer.setMouseNavEnabled(true);
         // enable the navigator
@@ -354,62 +159,8 @@ export class StillImageComponent implements OnInit, OnChanges, OnDestroy {
     viewer.open(tileSources);
   }
 
-  private _createSVGOverlay(
-    regionIri: string,
-    geometry: RegionGeometry,
-    aspectRatio: number,
-    regionLabel: string,
-    regionComment: string
-  ): void {
-    const viewer = this._assertOSDViewer();
-    const lineColor = geometry.lineColor;
-    const lineWidth = geometry.lineWidth;
-
-    const regEle: HTMLElement = this._renderer.createElement('div');
-    regEle.id = `region-overlay-${Math.random() * 10000}`;
-    regEle.className = 'region';
-    regEle.setAttribute('style', `outline: solid ${lineColor} ${lineWidth}px;`);
-
-    const diffX = geometry.points[1].x - geometry.points[0].x;
-    const diffY = geometry.points[1].y - geometry.points[0].y;
-
-    const loc = new OpenSeadragon.Rect(
-      Math.min(geometry.points[0].x, geometry.points[0].x + diffX),
-      Math.min(geometry.points[0].y, geometry.points[0].y + diffY),
-      Math.abs(diffX),
-      Math.abs(diffY * aspectRatio)
-    );
-
-    loc.y *= aspectRatio;
-
-    viewer
-      .addOverlay({
-        element: regEle,
-        location: loc,
-      })
-      .addHandler('canvas-click', event => {
-        this._regionService.highlightRegion((<any>event).originalTarget.dataset.regionIri);
-      });
-
-    this._regions[regionIri].push(regEle);
-
-    const comEle: HTMLElement = this._renderer.createElement('div');
-    comEle.className = 'annotation-tooltip';
-    comEle.innerHTML = `<strong>${regionLabel}</strong><br>${regionComment}`;
-    regEle.append(comEle);
-
-    regEle.addEventListener('mousemove', (event: MouseEvent) => {
-      comEle.setAttribute('style', `display: block; left: ${event.clientX}px; top: ${event.clientY}px`);
-    });
-    regEle.addEventListener('mouseleave', () => {
-      comEle.setAttribute('style', 'display: none');
-    });
-    regEle.dataset['regionIri'] = regionIri;
-  }
-
   private _loadImages() {
-    this._viewer?.close();
-    this.resetStatus();
+    this._assertOSDViewer().close();
 
     if (this.resource.properties[Constants.HasStillImageFileValue][0].type === Constants.StillImageFileValue) {
       this._loadInternalImages();
@@ -424,26 +175,20 @@ export class StillImageComponent implements OnInit, OnChanges, OnDestroy {
     const projectShort = this.resource.attachedToProject.split('/').pop();
     const assetId = this.imageFileValue?.filename.split('.')[0] || '';
 
-    if (!projectShort || !assetId) {
-      this._onFailedImageLoad();
-      return;
+    if (!projectShort) {
+      throw new AppError('Error with project shortcode');
     }
 
     this._rs
       .getIngestFileInfo(projectShort, assetId)
       .pipe(take(1))
-      .subscribe(
-        (fileInfo: FileInfo) => {
-          this.imageFileInfo = fileInfo;
-          if (this.failedToLoad) {
-            this._onSuccessAfterFailedImageLoad();
-          }
-          this._openInternalImages();
-        },
-        () => {
-          this._onFailedImageLoad();
+      .subscribe((fileInfo: FileInfo) => {
+        this.imageFileInfo = fileInfo;
+        if (this.failedToLoad) {
+          this._onSuccessAfterFailedImageLoad();
         }
-      );
+        this._openInternalImages();
+      });
   }
 
   private _loadExternalIIIF() {
@@ -454,20 +199,8 @@ export class StillImageComponent implements OnInit, OnChanges, OnDestroy {
           this._onSuccessAfterFailedImageLoad();
         }
         this._viewer?.open(iiif.infoJsonUrl);
-      } else {
-        this._onFailedImageLoad();
       }
     }
-  }
-
-  private _onFailedImageLoad() {
-    const viewer = this._assertOSDViewer();
-    this.failedToLoad = true;
-    viewer.setMouseNavEnabled(false);
-    viewer.navigator.element.style.display = 'none';
-    this.regionDrawMode = false;
-    this._viewer?.removeAllHandlers('open');
-    this._cd.markForCheck();
   }
 
   private _onSuccessAfterFailedImageLoad() {
@@ -476,51 +209,5 @@ export class StillImageComponent implements OnInit, OnChanges, OnDestroy {
     viewer.setMouseNavEnabled(true);
     viewer.navigator.element.style.display = 'block';
     this._regionService.imageIsLoaded();
-  }
-
-  private _renderRegions() {
-    let imageXOffset = 0; // see documentation in this.openImages() for the usage of imageXOffset
-
-    const stillImage = this.imageFileValue as ReadStillImageFileValue;
-    const aspectRatio = stillImage.dimY / stillImage.dimX;
-
-    const geometries = StillImageHelper.collectAndSortGeometries(this._regionService.regions, this._regions);
-
-    // render all geometries for this page
-    for (const geom of geometries) {
-      const geometry = geom.geometry;
-
-      const colorValues: ReadColorValue[] = geom.region.properties[Constants.HasColor] as ReadColorValue[];
-
-      // if the geometry has a color property, use that value as the color for the line
-      if (colorValues && colorValues.length) {
-        geometry.lineColor = colorValues[0].color;
-      }
-
-      const commentValue = geom.region.properties[Constants.HasComment]
-        ? geom.region.properties[Constants.HasComment][0].strval
-        : '';
-
-      if (!this.failedToLoad) {
-        this._createSVGOverlay(geom.region.id, geometry, aspectRatio, geom.region.label, commentValue || '');
-      }
-
-      imageXOffset++;
-    }
-  }
-
-  private _removeOverlays() {
-    for (const reg in this._regions) {
-      if (reg in this._regions) {
-        for (const pol of this._regions[reg]) {
-          if (pol instanceof HTMLElement) {
-            pol.remove();
-          }
-        }
-      }
-    }
-
-    this._regions = {};
-    this._assertOSDViewer().clearOverlays();
   }
 }
